Refactored imports and implemented caching for improved performance
Replaced include and url imports with re_path and path imports from django.urls in the urls.py files of the project and the mailer app. Also, updated the index.html template in the mailer app to use load static instead of load staticfiles. Furthermore, wrapped the IndexView class with the cache_page class in the mailer app's urls.py file to enable caching and reduce page load time on subsequent refreshes.